requirements.txt
alembic==0.8.5
blinker==1.4
dominate==2.2.0
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Moment==0.5.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
ForgeryPy==0.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
PyMySQL==0.7.2
python-editor==1.0
python-slugify==1.2.0
SQLAlchemy==1.0.12
Unidecode==0.4.19
visitor==0.1.2
Werkzeug==0.11.5
WTForms==2.1
